pyproject.toml
[project]
name = "ll-code-core"
version = "0.1.0"
description = "LL-Code Core - AI-powered coding assistant backend"
authors = [{ name = "Fromsko", email = "fromsko@users.noreply.github.com" }]
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "cryptography>=46.0.7",
    "fastapi>=0.136.0",
    "fastapi-users[sqlalchemy]>=15.0.5",
    "litellm>=1.83.10",
    "lxml>=6.1.0",
    "mcp>=1.27.0",
    "pydantic-settings>=2.14.0",
    "pyjwt>=2.12.1",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.26",
    "rich>=15.0.0",
    "sqlmodel>=0.0.38",
    "uvicorn[standard]>=0.45.0",
]

[project.scripts]
ll-start = "app.tui.launcher:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F","W", "UP"]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
